apps/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
openai>=1.10.0
pillow>=10.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
reportlab>=4.0.0
pypdf>=3.17.0
python-docx>=1.1.0
